Prevent objc name-collisions during run-time (macOS)
Add NETW_DELEGATE_NAME macro to easily change the name of the
Objective-C delegate used under macOS.
If netw (and thusly the delegate) is loaded multiple times in the
same process (e.g. multiple libraries being loaded, each of them
using their own version of netw) only one of the delegates is used
for all instances.
This breaks spectacularly when different versions of netw are used.